Define Fold operations using MonadUnliftIO
And move the FoldM "effect" monad to hs-bindgen from hs-bindgen-libclang
