+++ hs-bindgen-libclang/hs-bindgen-libclang.cabal
@@ -88,6 +88,7 @@ library
     , mtl         >= 2.2  && < 2.4
     , pretty-show >= 1.10 && < 1.11
     , text        >= 1.2  && < 2.2
+    , unliftio-core ^>=0.2.1.0
 
   if impl(ghc < 9.4)
     build-depends: data-array-byte >=0.1.0.1 && <0.2
