+++ hs-bindgen/hs-bindgen.cabal
@@ -66,6 +66,7 @@ library
       HsBindgen.Hs.AST
       HsBindgen.Backend.TH.Translation
       HsBindgen.Backend.PP.Render
+      HsBindgen.Eff
   other-modules:
       HsBindgen.Backend.PP
       HsBindgen.Backend.PP.Render.Internal
@@ -128,6 +129,7 @@ library
     , text               >= 1.2     && < 2.2
     , vec                >= 0.5     && < 0.6
     , some
+    , unliftio-core      ^>=0.2.1.0
 
 executable hs-bindgen
   import:         lang
