+++ hs-bindgen/src/HsBindgen/Imports.hs
@@ -12,6 +12,7 @@ import Control.Exception as X (Exception, throwIO, bracket)
 import Control.Monad as X (void, ap, forM_, guard)
 import Control.Monad.Identity as X (Identity (..))
 import Control.Monad.IO.Class as X (MonadIO (liftIO))
+import Control.Monad.IO.Unlift as X (MonadUnliftIO (withRunInIO))
 import Data.Bifunctor as X (Bifunctor (bimap, first, second))
 import Data.Coerce as X (coerce)
 import Data.Default as X (Default (def))
